+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 + Migrate from CircleCI + AppVeyor to GitHub Actions
 + Check datatype declaration (#15)
 + Check constructor declaration (#17)
++ Test desugar
++ Check declaration list (#19)
 
 # 0.0.4